systemd: do not fail 'stop' if service unit not installed
If a service unit configuration file doesn't exist on disk, 'systemctl stop'
fails with error code 5 [1]. To keep 'git-bundle-server web-server stop'
from failing when 'web-server start' was not called before it, do not cause
the program to fail when the error code indicates the service isn't
installed. Add a test covering this case.

[1] https://freedesktop.org/software/systemd/man/systemd.exec.html#Process%20Exit%20Codes

+++ internal/daemon/systemd.go
@@ -18,6 +18,8 @@ Type=simple
 ExecStart={{sq_escape .Program}}{{range .Arguments}} {{sq_escape .}}{{end}}
 `
 
+const SystemdUnitNotInstalledErrorCode int = 5
+
 type systemd struct {
 	user       common.UserProvider
 	cmdExec    common.CommandExecutor
@@ -107,7 +109,7 @@ func (s *systemd) Stop(label string) error {
 		return err
 	}
 
-	if exitCode != 0 {
+	if exitCode != 0 && exitCode != SystemdUnitNotInstalledErrorCode {
 		return fmt.Errorf("'systemctl stop' exited with status %d", exitCode)
 	}
